Add user update functionality and associated tests
This commit introduces the functionality for users to update their name and password. An `UpdateUserRequest` model has been added in the user model. A PATCH route `/current` has been created where users can make update requests. For validation, update rules have been added in the user validation file and associated tests to ensure the correctness of the new feature. The logs in the user services are also changed to debug level.