Add delete functionality to Contact
This commit introduces a 'delete' method to the contact controller to handle the deletion of contacts. An endpoint mapping to 'DELETE /:contactId' is added, which involves using the 'Delete' and 'HttpCode' decorators. The contact service is updated with a 'remove' function that communicates with the prisma service to remove the contact from the database. Corresponding tests for deleting contacts are also written in 'contact.spec.ts'.